pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "flask-jwt-extended>=4.7.1",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "python-jose>=3.4.0",
    "email-validator>=2.2.0",
]
